refactor: add shared hook for auth redirect
